@@ -65,19 +70,21 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
-
-        <spring-boot.version>1.5.10.RELEASE</spring-boot.version>
+        <java.version>21</java.version>
+        <spring-boot.version>3.5.0</spring-boot.version>
         <commons-collections4.version>4.4</commons-collections4.version>
-        <commons-lang3.version>3.12.0</commons-lang3.version>
+        <commons-lang3.version>3.14.0</commons-lang3.version>
         <mysql-connector-java.version>8.0.28</mysql-connector-java.version>
-        <druid.version>1.1.20</druid.version>
-        <baomidou-mybatis-plus.version>3.2.0</baomidou-mybatis-plus.version>
-        <baomidou-dynamic-datasource.versioin>2.5.7</baomidou-dynamic-datasource.versioin>
+        <druid.version>1.2.18</druid.version>
+        <baomidou-mybatis-plus.version>3.5.5</baomidou-mybatis-plus.version>
+        <baomidou-dynamic-datasource.versioin>4.2.0</baomidou-dynamic-datasource.versioin>
         <java-uuid-generator.version>3.1.4</java-uuid-generator.version>
         <fastjson.version>1.2.83</fastjson.version>
-        <groovy.version>2.3.7</groovy.version>
-        <guava.version>31.0.1-jre</guava.version>
+        <groovy.version>4.0.22</groovy.version>
+        <guava.version>32.1.3-jre</guava.version>
+        <javax.annotation-api.version>1.3.2</javax.annotation-api.version>
+        <jakarta.annotation-api.version>2.1.1</jakarta.annotation-api.version>
+        <junit.version>4.13.2</junit.version>
     </properties>
 
     <dependencyManagement>
@@ -176,8 +193,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>21</source>
+                    <target>21</target>
                     <encoding>UTF-8</encoding>
                 </configuration>
             </plugin>
