@@ -11,7 +11,6 @@ import com.didiglobal.turbo.engine.model.InstanceData;
 import com.didiglobal.turbo.engine.spi.HookService;
 import com.didiglobal.turbo.engine.util.FlowModelUtil;
 import com.didiglobal.turbo.engine.util.InstanceDataUtil;
-import com.google.common.collect.Lists;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.MapUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -28,6 +27,8 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 public class ExclusiveGatewayExecutor extends ElementExecutor implements InitializingBean {
@@ -76,20 +77,23 @@ public class ExclusiveGatewayExecutor extends ElementExecutor implements Initial
     }
 
     private Map<String, InstanceData> getHookInfoValueMap(String flowInstanceId, String hookInfoParam, String nodeKey, String nodeInstanceId) {
-        List<InstanceData> dataList = Lists.newArrayList();
-        for (HookService service : hookServices) {
-            try {
-                List<InstanceData> list = service.invoke(flowInstanceId, hookInfoParam, nodeKey, nodeInstanceId);
-                if (CollectionUtils.isEmpty(list)) {
-                    LOGGER.warn("hook service invoke result is empty, serviceName={}, flowInstanceId={}, hookInfoParam={}",
-                        service.getClass().getName(), flowInstanceId, hookInfoParam);
-                }
-                dataList.addAll(list);
-            } catch (Exception e) {
-                LOGGER.warn("hook service invoke fail, serviceName={}, flowInstanceId={}, hookInfoParam={}",
-                    service.getClass().getName(), flowInstanceId, hookInfoParam);
-            }
-        }
+        List<InstanceData> dataList = hookServices.stream()
+                .flatMap(service -> {
+                    try {
+                        List<InstanceData> list = service.invoke(flowInstanceId, hookInfoParam, nodeKey, nodeInstanceId);
+                        if (CollectionUtils.isEmpty(list)) {
+                            LOGGER.warn("hook service invoke result is empty, serviceName={}, flowInstanceId={}, hookInfoParam={}",
+                                    service.getClass().getName(), flowInstanceId, hookInfoParam);
+                            return Stream.empty();
+                        }
+                        return list.stream();
+                    } catch (Exception e) {
+                        LOGGER.warn("hook service invoke fail, serviceName={}, flowInstanceId={}, hookInfoParam={}",
+                                service.getClass().getName(), flowInstanceId, hookInfoParam);
+                        return Stream.empty();
+                    }
+                })
+                .collect(Collectors.toList());
         return InstanceDataUtil.getInstanceDataMap(dataList);
     }
 
